pyproject.toml
[tool.poetry]
name = "ruff2bitbucket"
version = "0.1.0"
description = "Send your ruff results to bitbucket as annotations."
authors = ["Steven Van Ingelgem <steven@vaningelgem.be>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
ruff = "*"
requests = "*"

[tool.poetry.group.dev.dependencies]
pytest-mock = "*"
pytest-cov = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
test = "pytest"
ruff2bitbucket = "ruff2bitbucket.__main__:main"

[tool.ruff]
# https://docs.astral.sh/ruff/rules
select = [
    "E",  # pycodestyle/flake8
    "W",  # "
    "F",  # pyflakes
    "C",  # mccabe
    "I",  # isort
    "B",  # flake8-bugbear
    "Q",  # flake8-quotes
    "N",  # pep8-naming
    "ANN",  # flake8-annotations
    "A",  # flake8-builtins
    #    "COM",  # flake8-comma
    "C4",  # flake8-comprehensions
    "ISC",  # flake8-implicit-str-concat
    "ICN",  # flake8-import-conventions
    "G",  # flake8-logging-format
    ##    "INP",  # flake8-no-pep420
    #    "PIE",  # flake8-pie
    "T20",  # flake8-print
    #    "PYI",  # flake8-pyi
    "PT",  # flake8-pytest-style
    "RET",  # flake8-return
    #    "SLF",  # flake8-self
    "SIM",  # flake8-simplify
    #    "TID",  # flake8-tidy-imports
    #    "TCH",  # flake8-type-checking
    "ARG",  # flake8-unused-arguments
    #    "PTH",  # flake8-use-pathlib
    #    "ERA",  # eradicate
    #    "PD",  # pandas-vet
    #    "PL",  # pylint
    #    "FLY",  # flynt
    #    "NPY",  # NumPy-specific rules
    #    "AIR",  # Airflow
    #    "PERF",  # Perflint
    ##    "FURB",  # refurb
    ##    "LOG",  # flake8-logging
    "RUF",  # Ruff-specific rules
]
ignore = [
    "ANN101",
    "ANN102",
    "ISC001",
]
fixable = ["ALL"]
line-length = 120
target-version = "py39"
fix = true

[tool.ruff.isort]
case-sensitive = false


[tool.pytest.ini_options]
testpaths = "tests"

log_cli = true
